Multi-Object (MO) UCX backend implementation (rebased over code restructuring PR) (#58)
* backend/ucx_mo: implement multi-object

1. backend/ucx: Move CUDA declarations to cpp-file

Isolate header files from compile-time dependencies
(may not be required if global config.h is distributed
along with header files)
2. ucx/test: Fix progress (when p-thread is disabled)

Progress may be required for both ucx1 and ucx2
in different stages

3. Add a new UCX-based backend that allows associating NIXL
logical "devices" with different UCX workers.
The primary motivation is that UCX v1.18 doesn't
support more than one GPGPU per UCX context.

NOTE: This is expected to be fixed in UCX v1.19
so this backend might be viewed as a workaround
unless other uses will be found.

Signed-off-by: Artem Y. Polyakov <artemp@nvidia.com>