Extend UCX engine with multi worker support (#257)
* Cleanup UCX progress thread

Small cleanups in progress thread that simultaneously make it more robust,
efficient and less likely to generate false-positives with Thread
sanitizer:

 - Use atomics instead of volatile for simple cases.

 - Use conditional variable to allow caller to sleep while thread is being
 initialized instead of busy-waiting.

 - Refactor iteration amount constant to be static constexpr instead of a
 class member.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

* UCX data structures refactoring

 - Move methods that are not actually related to particular worker to
 classes they belong to.

 - Improve memory safety by using smart pointers across the whole code. Use
 strictest smart pointer that fits the use-case (i.e. unique_ptr where
 possible).

 - Don't arbitrary create temporary copies of objects where it is not
 strictly necessary. Access the instances via an iterator to the container
 that owns them.

 - Properly initialize/cleanup nixlUcpEp instances via class ctor/dtor
 instead of member methods that can be called zero or multiple times per
 object instead of exactly once.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

* Extend UCX backend to support multiple workers

Configure number of workers by setting new agent custom parameter
'num_workers'.

Spawn multiple workers in UCX backend and establish connection on all of
them but decide which instance to use for data-path operations based on
caller thread id hash to allow equal spread of tasks over multiple workers
in a way that is transparent to the user.

For the purposes of generating metadata the first worker is always chosen.

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>

---------

Signed-off-by: Vlad Buslov <vladbu@nvidia.com>