Add NIXL Telemetry infrastructure (#562)
Add NIXL Telemetry infrastructure

This commit introduces infrastructure and basic agent telemetry collection.
Telemetry data is written to a shared memory and can be read
and processd by other applications. A simple reader is added.
This commit adds a cyclic buffer and a Thread executer that
can be used across NIXL for other use cases.
---------

Signed-off-by: Rafi Wiener <rwiener@nvidia.com>
Co-authored-by: Rafi Wiener <rwiener@nvidia.com>
Co-authored-by: Moein Khazraee <moein@nvidia.com>